# server/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Loventia API
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: Loventia API Team
    url: https://loventia.example.com
    email: api@loventia.example.com
  description: |
    OpenAPI specification for the Loventia dating application.

    This spec is aligned with the current server/src/app.js routing we just test-drove
    with PowerShell:

      - Primary auth endpoint: POST /api/auth/login
      - Legacy/fallback auth endpoint: POST /api/users/login
      - Refresh: POST /api/auth/refresh
      - Self: GET /api/auth/me and GET /api/users/me and GET /api/me
      - Reset-password flow: POST /api/auth/forgot-password, POST /api/auth/reset-password
      - Private diagnostics: GET /api/auth/private/ping, GET /api/auth/private/no-me
      - User photo management (the things we just tested): 
          * DELETE /api/users/{userId}/photos/{slot}
          * DELETE /api/users/{userId}/photos?path=...
          * PUT    /api/users/{userId}/photos/reorder
          * POST   /api/users/{userId}/set-avatar
      - Billing endpoints (Stripe webhook at /webhooks/stripe and legacy /api/payment/stripe-webhook)
      - Admin CSV exports
      - Health/readiness endpoints

    IMPORTANT:
      - We **document both** logins on purpose:
          1) /api/auth/login  (main, preferred)
          2) /api/users/login (legacy/fallback)
      - This matches the README change you will add in server/README.md
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: villehermaala1981@gmail.com
        password:
          type: string
          example: Paavali1981
    AuthLoginResponse:
      type: object
      properties:
        message:
          type: string
          example: Login successful.
        user:
          $ref: '#/components/schemas/UserSelf'
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          example: 7200
        # OpenAPI 3.1: use union type instead of nullable
        refreshExpiresIn:
          type: [integer, 'null']
    AuthRefreshRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    AuthRefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
        user:
          $ref: '#/components/schemas/UserSelf'
...
    UserSelf:
      type: object
      description: >
        Minimal user shape returned by /api/auth/me and /api/users/me in our PS tests.
        Includes entitlements, visibility and premium flags.
      properties:
        id:
          type: string
          example: 68dab8a761a412065159ff4d
        _id:
          type: string
          example: 68dab8a761a412065159ff4d
        email:
          type: string
          example: villehermaala1981@gmail.com
        username:
          type: string
          example: Ville1981
        role:
          type: string
          example: user
        premium:
          type: boolean
          example: true
        isPremium:
          type: boolean
          example: true
        entitlements:
          type: object
          properties:
            tier:
              type: string
              example: premium
            # OpenAPI 3.1 union types (replace nullable)
            since:
              type: [string, 'null']
            until:
              type: [string, 'null']
            features:
              type: object
              additionalProperties: true
            quotas:
              type: object
              additionalProperties: true
        visibility:
          type: object
          properties:
            isHidden:
              type: boolean
              example: false
            # OpenAPI 3.1 union type (replace nullable)
            hiddenUntil:
              type: [string, 'null']
            resumeOnLogin:
              type: boolean
              example: true
        profilePicture:
          type: string
          example: /uploads/1759343842730-848316731.gif
        photos:
          type: array
          items:
            type: string
            example: /uploads/1759343842730-848316731.gif
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProfileCard:
      type: object
      description: Public information displayed in Discover listings.
      properties:
        id:
          type: string
          example: 68dab8a761a412065159ff4d
        username:
          type: string
          example: Ville1981
        age:
          type: integer
          example: 43
        gender:
          type: string
          example: male
        bio:
          type: string
          example: Coffee, synthwave and long runs by the sea.
        distanceKm:
          type: number
          example: 12.5
        profilePicture:
          type: string
          example: /uploads/1759343842730-848316731.gif
        photos:
          type: array
          items:
            type: string
          example:
            - /uploads/p1.jpg
            - /uploads/p2.jpg
        location:
          type: object
          properties:
            lat:
              type: number
              example: 60.1699
            lng:
              type: number
              example: 24.9384
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        limit:
          type: integer
          minimum: 1
          example: 20
        total:
          type: integer
          minimum: 0
          example: 237
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
...
paths:
  /api/discover:
    get:
      tags: [Discover]
      summary: Discover profiles
      description: Returns a paginated list of public profile cards.
      operationId: discover_list
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: minAge
          schema: { type: integer, minimum: 18 }
        - in: query
          name: maxAge
          schema: { type: integer, minimum: 18 }
        - in: query
          name: gender
          schema: { type: string }
        - in: query
          name: distanceKm
          schema: { type: number }
        - in: query
          name: lat
          schema: { type: number }
        - in: query
          name: lng
          schema: { type: number }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProfileCard'
                  meta:
                    $ref: '#/components/schemas/Pagination'
        '400': { description: Invalid query parameters }
        '401': { description: Unauthorized }
...
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login (primary)
      description: >
        Primary login endpoint for Loventia. FE calls this FIRST.
        If it returns 404/405, FE may try /api/users/login (legacy).
      operationId: auth_login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '400':
          description: Invalid credentials or missing payload
...
  /api/auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: >
        Matches the PS call: POST /api/auth/refresh { "refreshToken": "<token>" }
        Returns a fresh access token and user payload.
      operationId: auth_refresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRefreshResponse'
        '401':
          description: Invalid or expired refresh token
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Get current user (auth router)
      description: Return the authenticated user from the auth router.
      operationId: auth_me
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user (from auth router)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSelf'
        '401':
          description: Unauthorized
  /api/me:
    get:
      tags: [Authentication]
      summary: Get current user (central route)
      description: >
        Centralized self-route. We confirmed with PS that this returns
        the same normalized shape as /api/auth/me.
      operationId: me_central
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user (centralized)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSelf'
        '401':
          description: Unauthorized
...
  /webhooks/stripe:
    post:
      tags: [Billing]
      summary: Stripe webhook (raw body)
      description: >
        Real webhook endpoint as mounted in server/src/app.js:
        app.post('/webhooks/stripe', express.raw({ type: 'application/json' }), ...)
        Requires Stripe-Signature header. Raw body is required.
      operationId: billing_webhook_stripe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Acknowledged
        '400':
          description: Invalid signature or payload
...
  /__routes:
    get:
      tags: [Diagnostics]
      summary: List app-level routes (dev-only)
      description: Returns app-level mounted routes (dev-only).
      operationId: diagnostics_routes
      responses:
        '200':
          description: List of routes
        '401':
          description: Unauthorized
  /__routes_api:
    get:
      tags: [Diagnostics]
      summary: List /api router routes (dev-only)
      description: Returns API router routes (dev-only).
      operationId: diagnostics_routes_api
      responses:
        '200':
          description: List of /api routes
        '401':
          description: Unauthorized
  /__routes_api_full:
    get:
      tags: [Diagnostics]
      summary: List /api router routes (detailed)
      description: Returns detailed list of API routes (dev-only).
      operationId: diagnostics_routes_api_full
      responses:
        '200':
          description: Detailed list of /api routes
        '401':
          description: Unauthorized
security:
  - bearerAuth: []